app/requirements.txt
Flask==2.3.2
google-cloud-pubsub==2.19.0